Add typechecking that cover_image_url exists and is a string

+++ src/routes/+page.server.ts
@@ -44,13 +44,20 @@ export const actions = {
 	delete: async ({ request }) => {
 		const data = await request.formData();
 		const id = data.get("id");
-
+		const cover_image_url = String(data.get("cover_image_url"));
 		try {
+			if (!cover_image_url) {
+				return;
+			}
+			await supabase.storage
+				.from("cover_images")
+				.remove([cover_image_url]);
 			const { data: recordToRemove, error: deleteError } = await supabase
 				.from("media")
 				.delete()
 				.eq("id", id);
 
+			// also delete the stored cover image
 			if (deleteError) {
 				throw new Error(deleteError.message);
 			}
@@ -65,12 +72,28 @@ export const actions = {
 		const id = data.get("id");
 		const title = data.get("title");
 		const media_type = data.get("media_type");
+		// new cover image the user wants
 		const cover_image = data.get("cover_image");
+		// old url of what is currently rendered
+		const cover_image_url = String(data.get("cover_image_url"));
 
-		if (!cover_image || !(cover_image instanceof Blob)) {
+		if (!cover_image_url) {
 			return;
 		}
+		let new_cover_image_url = cover_image_url;
+		if (cover_image && cover_image instanceof Blob) {
+			await supabase.storage
+				.from("cover_images")
+				.remove([cover_image_url]);
+			const new_cover_image_url = uploadImage(cover_image);
+		}
 
-		const cover_image_url = uploadImage(cover_image);
+		// remember to delete the old cover_image from the bucket
+
+		const { data: recordToUpdate, error } = await supabase
+			.from("media")
+			.update({ title, media_type, cover_image_url: new_cover_image_url })
+			.eq("id", id)
+			.select();
 	},
 };
